requirements.txt
streamlit
mysql-connector-python
cryptography
psycopg2-binary
python-dotenv
pycryptodome